Remove old --deps option on creating an app
It seems installing dependencies has become a default behavior, and the `--deps` option becoming deprecated (replaced with `--no-deps`). Therefore having this option can lead to confusion. It could be replaced with the `--no-deps` option, but that's maybe out of scope for this situation, as not all options are documented here anyway and for those beginning to use Amber it would not be relevant information.